Changed id to Long

+++ src/main/java/ca/sfu/delta/models/DistributionEntry.java
@@ -7,18 +7,19 @@ import javax.persistence.Id;
 
 @Entity
 public class DistributionEntry {
-	@Id @GeneratedValue(strategy=GenerationType.AUTO)
-	private Integer id;
+	@Id
+	@GeneratedValue(strategy=GenerationType.AUTO)
+	private Long id;
 
 	private String name;
 
 	private String email;
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
